+++ disperser/apiserver/validation.go
@@ -3,6 +3,7 @@ package apiserver
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/Layr-Labs/eigenda/core"
 	"github.com/Layr-Labs/eigenda/encoding"
@@ -134,5 +135,63 @@ func (s *DispersalServer) checkRateLimitsAndAddRates(ctx context.Context, blob *
 		blob.RequestHeader.SecurityParams[i].QuorumRate = accountRates.Throughput
 	}
 	return nil
+
+}
+
+func (s *DispersalServer) getAccountRate(origin, authenticatedAddress string, quorumID core.QuorumID) (*PerUserRateInfo, string, error) {
+	unauthRates, ok := s.rateConfig.QuorumRateInfos[quorumID]
+	if !ok {
+		return nil, "", fmt.Errorf("no configured rate exists for quorum %d", quorumID)
+	}
+
+	rates := &PerUserRateInfo{
+		Throughput: unauthRates.PerUserUnauthThroughput,
+		BlobRate:   unauthRates.PerUserUnauthBlobRate,
+	}
+
+	// Check if the address is in the allowlist
+	if len(authenticatedAddress) > 0 {
+		quorumRates, ok := s.rateConfig.Allowlist[authenticatedAddress]
+		if ok {
+			rateInfo, ok := quorumRates[quorumID]
+			if ok {
+				key := "address:" + authenticatedAddress
+				if rateInfo.Throughput > 0 {
+					rates.Throughput = rateInfo.Throughput
+				}
+				if rateInfo.BlobRate > 0 {
+					rates.BlobRate = rateInfo.BlobRate
+				}
+				return rates, key, nil
+			}
+		}
+	}
+
+	// Check if the origin is in the allowlist
+
+	key := "ip:" + origin
+
+	for account, rateInfoByQuorum := range s.rateConfig.Allowlist {
+		if !strings.Contains(origin, account) {
+			continue
+		}
+
+		rateInfo, ok := rateInfoByQuorum[quorumID]
+		if !ok {
+			break
+		}
+
+		if rateInfo.Throughput > 0 {
+			rates.Throughput = rateInfo.Throughput
+		}
+
+		if rateInfo.BlobRate > 0 {
+			rates.BlobRate = rateInfo.BlobRate
+		}
+
+		break
+	}
+
+	return rates, key, nil
 
 }
